Se agrega estructura y factories para app y db
