+++ package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/ngrosso/coderHouseNode2023#readme",
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "commander": "^10.0.1",
     "connect-mongo": "^5.0.0",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.3",
@@ -28,6 +29,7 @@
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.0.4",
     "mongoose-paginate-v2": "^1.7.1",
+    "shelljs": "^0.8.5",
     "zod": "^3.21.4"
   },
   "devDependencies": {
