+++ src/presentation/factories/appFactory.js
@@ -1,0 +1,15 @@
+import AppExpress from "../application/appExpress.js";
+
+class AppFactory
+{
+    static create(appType)
+    {
+        const apps = new Map();
+        apps.set('AppExpress', AppExpress);
+
+        const app = apps.get(appType);
+        return new app();
+    }
+}
+
+export default AppFactory;
