+++ src/presentation/middlewares/admin.middleware.js
@@ -1,0 +1,21 @@
+import jwt from "jsonwebtoken";
+import config from "../../config/index.js";
+
+const roleAuth = (req, res, next) => {
+  const authHeader = req.headers.authorization ?? `Bearer ${req.cookies.accessToken}`;
+  if (!authHeader) {
+    return res.status(401).send({ message: 'Empty authentication header!' })
+  }
+
+  const token = authHeader.split(' ')[1]; // Bearer tokenString
+
+  jwt.verify(token, config.jwtPrivateKey, (error, credentials) => {
+    if (error || !credentials.user.admin) return res.status(403).send({ error: 'Admin Authentication error' });
+
+    req.email = credentials.email;
+    next();
+  });
+
+}
+
+export default roleAuth;
