+++ src/presentation/middlewares/errorHandler.middleware.js
@@ -1,0 +1,5 @@
+const errorHandler = (err, req, res, next) => {
+  //TODO: Implementar
+}
+
+export default errorHandler;
