--- src/utils/auth.js
@@ -1,14 +1,0 @@
-import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
-
-export const createHash = async (password) => {
-  return await bcrypt.hash(password, 10)
-}
-
-export const isValidPassword = async (password, passwordHash) => {
-  return await bcrypt.compare(password, passwordHash);
-}
-
-export const generateToken = async (user) => {
-  return await jwt.sign({ user: { ...user, password: undefined } }, process.env.PRIVATE_KEY, { expiresIn: '60m' });
-}
